#515 Sequential Activities Summary Switcher
